LakeShark 2.2.1
- Install writes the complete firmware, on a new board or over an existing one.
- Boots to the interface whatever state the storage partition is in.
- Sub-GHz scan and learn, SX1262 FSK transmit and a capture archive.
- GFSK and POCSAG in LORA LABS, with launch animations restored.
- One .sub writer, universal record and replay, and readable menus.
- FM touch routing fixed, and the FM spectrum leaves the radio alone.
- SUB-GHZ keeps its landscape rows and the top bar reaches the glass.
- P25 profile format version 2 carries site coordinates.

+++ components/lakeshark/core/ls_ctl.c
@@ -455,7 +455,8 @@ static int cmd_p25enc(int argc, char **argv)
                    nm ? nm : "?",
                    (unsigned)P25.p25_enc_tg[i].kid);
             if (P25.p25_enc_tg[i].skip_remaining_ms > 0)
-                printf("  skip=%ds", P25.p25_enc_tg[i].skip_remaining_ms / 1000);
+                printf("  skip=%ds",
+                       (int)(P25.p25_enc_tg[i].skip_remaining_ms / 1000));
             printf("\n");
         }
     }
